Retropie #7/8: VLC (#169)
VLC port, split from PR https://github.com/qnx-ports/build-files/pull/58
Retropie.

This is an adapted version of a previous port by Darcy, made on GitLab.
The supported version is 3.0.21. The companion GitHub Repo can be found
at https://github.com/qnx-ports/vlc.

**On the Structure of these Files**
Please note that the build-files structure for this PR is set up
differently from normal. This is a result of three design constraints:
1) VLC's build process is being 'hijacked' by our QNX makefiles (This is
kept from the internal port available on GitLab). As a result, the build
tree needs to be in a qnx subfolder of the source tree.
2) RetroPie's build script requires an install_rpie option that
generates a staging folder and artifacts in the build-files directory,
such that it can be copied into RetroPie's staging setup. This follows
the convention of all other dependencies.
3) Our porting process requires that build files be split into a
separate repo from the source code.

This results in the structure we have here. There are two sets of
makefiles - slightly modified ones from the original port which exist in
the ports/vlc/qnx subfolder and the standard set in ports/vlc. The
standard files act as a 'wrapper', copying the vlc/qnx folder into the
source tree and correctly running the build process. It then extracts
the relevant artifacts from their locations in the build tree and places
them into the appropriate nto-* folders in build-files/ports/vlc.

This is unideal, but after extensive testing it seems this is the only
method that works and satisfied our porting process' requirement of
being placed here and RetroPie's requirement of having a set of build
files and artifacts available here.

**On the Workflow**
Currently, vlc's workflow seems to fail due to two reasons: autotools
being incompatible in the testing docker image, and missing qnx video
packages. Due to lack of permissions in the workflows, this seems to be
unfixable for now.
Due to this, I'm removing the workflow file.