RetroPie #8/8: RetroArch, EmulationStation, OpenTTD, Integration and Scripts (#58)
This PR is the final PR in the **RetroPie/RetroQNX effort**. While other
PRs encompassed single dependencies of the project, this PR is focused
on the core content, visuals, and engines that the demo runs on. It also
includes full documentation, installation and uninstall scripts, and
attribution.

There are 3 main segments of files to review: 
1. Documentation, including licensing and review. This is all the
README.md files in RetroPie and its subfolders, as well as
SCRIPTINFO.md. _i.e. is licensing satisfied & documentation sufficient
in your opinion?_
2. Installation & Build Process. This is the .sh scripts contained in
the RetroPie subfolder, and checking that the instructions in README.md
are sufficient. _i.e. does this actually work on your machine as well?_
3. Individual Files. The rest of the content, which is mostly layout,
configs, toolchain files, and the individual build files of the sub
parts. _i.e. is everything here and does it follow our process to the
best of its ability in your opinion?_

**For those in the same office as me:**
- I have a spare v1 demo SD card at my desk, alongside my development
machine which has a copy. Otherwise, you can follow the standard build &
installation process, though the build process can take multiple hours.
You'll need ~25 GB of space on the host machine to compile, and ~2GB of
space on your target to install. (Plus room for save files - ~4GB
recommended)
- I am also interested in running a networking "stress test" to
determine the limits of a pi when running OpenTTD-Server. This would
help determine how feasible a multiplayer hackathon demo could be. If
anyone is interested, let me know.

**Deviations from the Process**:
libretro-cores contains a patch file which is applied to each core. This
is because it isn't actually modifying anything about the cores
themselves, but rather the libretro provided Makefile, changing the std
to c++11. Since this is a provided file to *all* libretro cores,
individually changing each one via upstream is infeasible, particularly
as some (not included in this release) actually do need code changes.
Forking each would create a lot of essentially unchanged repos for us to
keep up with. And providing a patch file also means users can try to
compile more cores at their leisure.

**KNOWN ISSUES**
__OpenTTD is slow to boot__: This only occurs when booting through the
RetroQNX menu (EmulationStation). Load time spikes to ~ 1 minute, where
it is normally ~ 5 seconds.
__Controller doesn't work__: Likely not supported by usb-to-screen, as
the data-driven portions are upcoming and HID input is iffy at the
moment. Please message me the controller info - I'll try to include it
in v3 of that project, which is coming up soon. If your controller does
not work, please use keyboard mappings.
__Workflows__: Emulation Station and OpenTTD require a missing swc
package each, and thus do not yet work. I'm keeping the workflows for
now in case a solution is found.

**ASSOCIATED PRs:**
1/8 nanosvg: #120 
2/8 rapidjson: #121 
3/8 pugixml: #122 
4/8 FreeImage: #123 
5/8 SDL 2: #128
6/8 lua: #129
7/8 VLC: #169
8/8 Core: #58

**ASSOCIATED REPOS**
The following on qnx-ports: vlc, EmulationStation, SDL, rapidjson,
FreeImage, pugixml, nanosvg, lua, muslflt, openttd, RetroArch
The usb-to-screen qnx demo https://gitlab.com/qnx/projects/usb-to-screen
The qnx inspired theme es-theme-qnx:
https://github.com/qnx-ports/es-theme-qnx

From RetroArch Upstream, https://github.com/libretro/retroarch-assets,
https://github.com/libretro/libretro-core-info.
From Upstream as Content,
https://github.com/libretro/libretro-samples.git,
https://github.com/libretro/libretro-2048,
https://github.com/Javanaise/mrboom-libretro.git,
https://github.com/Jakz/retro8.git,
https://github.com/jtothebell/fake-08.git