protobuf: Set language specific flags directly for C/CXX, added ZLIB location (#268)
For some reason, CMake still supplies C objects with `CXXFLAGS` which
results in the following error:
```
/opt/qnx/qnx800/target/qnx/usr/include/c++/v1/cstddef:46:5: error: #error <cstddef> tried including <stddef.h> but didn't find libc++'s <stddef.h> header. This usually means that your header search paths are not configured properly. The header search paths should contain the C++ Standard Library headers before any C Standard Library, and you are probably using compiler flags that make that not be the case.
   46 | #   error <cstddef> tried including <stddef.h> but didn't find libc++'s <stddef.h> header. \
      |     ^~~~~
```
To mitigate this, a proper distinction between `CFLAGS` and `CXXFLAGS`
is necessary.

Additionally, on some systems CMake picks up local ZLIB, instead of
using the one in the SDP.

Both of those issues are addressed in this PR. Additionally, it adds
information on how to deal with missing `librt` in the `README.md` of
protobuf.